NativeExport attribute; Dll added; Started working on generating script
