+++ Sources/Utils/Utils/ExportsGenerator.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Utils
+{
+    public static class ExportsGenerator
+    {
+        public static void Do()
+        {
+        }
+    }
+}
