+++ Sources/Utils/Utils/Utils.csproj
@@ -5,4 +5,8 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\AngouriMath\AngouriMath.csproj" />
+  </ItemGroup>
+
 </Project>
